refactor(netcdf): move sel into engines/selection.py (STR-1)
Relocate sel (band selection by coordinate value along a tracked band
dim) out of netcdf.py into Selection.sel (issue #615). NetCDF.sel becomes
a thin façade delegating to self.selection.sel; signature, behaviour, and
return type are unchanged. The engine reaches the variable's band-dim
metadata and helpers (_iloc / _preserve_netcdf_metadata /
_derive_primary_band_view) through self._ds, and builds the result via
Dataset.create_from_array exactly as before.
